third_party/WebKit/LayoutTests/fast/css/specificity-not-and-default-ns.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
  @namespace "http://www.w3.org/1999/xhtml";

  #t1:not(.nomatch) { color: green }
  #t1 { color: red }
</style>
<div id="t1">This text should be green</div>
<script>
  test(() => assert_equals(getComputedStyle(t1).color, "rgb(0, 128, 0)"),
      "Selectors inside :not contributes specificity.");
</script>
